Move most of the CSV logic to `TableMixin`.
